+++ amd/src/delta_recorder.js
@@ -17,6 +17,7 @@
  * @module     tiny_authory_tech/delta_recorder
  * @category   TinyMCE Editor
  * @copyright  2026 SEPTUM QA <info@authory.tech>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
  */
 
 import {call} from 'core/ajax';
@@ -702,6 +703,8 @@ export const register = (editor, interval, userId, hasApiKey, MODULES, Rubrics, 
                 getString('pluginname', 'mod_quiz'),
                 getString('pluginname', 'mod_lesson'),
                 getString('description', 'tiny_authory_tech'),
+                getString('pdfannotation', 'tiny_authory_tech'),
+                getString('page', 'tiny_authory_tech'),
             ]).then(function(strings) {
                 return localStorage.setItem('sbTitle', JSON.stringify(strings));
             }).catch(error => window.console.error(error));